update eclipse project

+++ Common/Inc/shared.h
@@ -68,7 +68,7 @@ extern struct shared_var shared_var;
 extern struct params params;
 extern struct cisCals cisCals;
 extern int32_t imageData[CIS_PIXELS_NB];
-extern int32_t cisData[CIS_ADC_BUFF_SIZE * 3];
+extern int16_t cisData[CIS_ADC_BUFF_SIZE * 3];
 extern int32_t cisDataCpy[CIS_ADC_BUFF_SIZE * 3];
 
 extern int params_size;
