+++ Common/Src/shared.c
@@ -28,7 +28,7 @@ __attribute__ ((section(".imageData")))
 int32_t imageData[CIS_PIXELS_NB] = {0};
 
 __attribute__ ((section(".cisData")))
-int32_t cisData[CIS_ADC_BUFF_SIZE * 3] = {0};
+int16_t cisData[CIS_ADC_BUFF_SIZE * 3] = {0};
 
 __attribute__ ((section(".cisDataCpy")))
 int32_t cisDataCpy[CIS_ADC_BUFF_SIZE * 3] = {0};
